+++ src/picopie/__init__.py
@@ -78,4 +78,4 @@ __all__ = [
     "voxel_size",
 ]
 
-__version__ = "0.6.0"
+__version__ = "0.7.0"